backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
google-cloud-bigquery==3.11.4
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
requests==2.31.0
gunicorn==21.2.0
bcrypt==4.0.1
anthropic==0.7.8
python-dotenv==1.0.0
uuid
datetime
jsonschema==4.19.1
werkzeug==2.3.7
